requirements.txt
functions-framework
numpy
pygltflib
google-cloud-storage